+++ app/services/bitsy/check_payment_depots_transactions.rb
@@ -2,10 +2,10 @@ module Bitsy
   class CheckPaymentDepotsTransactions
 
     include LightService::Action
-    expects :latest_block
+    expects :latest_block, :payment_depots
 
     executed do |ctx|
-      PaymentDepot.for_manual_checking.find_each do |payment_depot|
+      ctx.payment_depots.find_each do |payment_depot|
         CheckPaymentDepotTransactions.execute(
           payment_depot: payment_depot,
           latest_block: ctx.latest_block,