+++ app/services/bitsy/fetch_payment_depots_for_manual_checking.rb
@@ -1,0 +1,12 @@
+module Bitsy
+  class FetchPaymentDepotsForManualChecking
+
+    include LightService::Action
+    promises :payment_depots
+
+    executed do |ctx|
+      ctx.payment_depots = PaymentDepot.for_manual_checking
+    end
+
+  end
+end