+++ spec/services/check_for_payments_spec.rb
@@ -5,7 +5,8 @@ module Bitsy
 
     it "executes actions" do
       actions = [
-        GetLatestBlock,
+        FetchPaymentDepotsForManualChecking,
+        GetLatestBlockIfPaymentDepots,
         CheckPaymentDepotsTransactions,
       ]